rename rules to lower case and add clear script to clean lib folder
