+++ package.json
@@ -4,6 +4,7 @@
   "description": "SonarTS rules for TSLint",
   "main": "tslint-sonarts.json",
   "scripts": {
+    "clear": "rm -rf lib/*",
     "prepublish": "tsc -p .",
     "build": "tsc -p . && yarn build-cfg && webpack",
     "build-cfg": "cp src/tools/cfg_viewer/index.html lib/tools/cfg_viewer/",
