+++ src/rules/noAllDuplicatedBranchesRule.ts
