+++ src/rules/noEmptyDestructuringRule.ts
