+++ src/rules/noIdenticalConditionsRule.ts
