+++ src/rules/noIgnoredReturnRule.ts
