+++ src/rules/noSelfAssignmentRule.ts
