+++ src/rules/noUseOfEmptyReturnValueRule.ts
