+++ tests/rules/noAllDuplicatedBranchesRule/noAllDuplicatedBranchesRule.lint.ts
