+++ tests/rules/noAllDuplicatedBranchesRule/noAllDuplicatedBranchesRule.test.ts
@@ -17,10 +17,10 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-import { Rule } from "../../../src/rules/NoUseOfEmptyReturnValueRule";
+import { Rule } from "../../../src/rules/noAllDuplicatedBranchesRule";
 import runRule from "../../runRule";
 
-it("test", () => {
-  const result = runRule(Rule, __filename);
-  expect(result.actualErrors).toEqual(result.expectedErrors);
+it("raises error", () => {
+  const { actualErrors, expectedErrors } = runRule(Rule, __filename);
+  expect(actualErrors).toEqual(expectedErrors);
 });
