+++ tests/rules/noEmptyDestructuringRule/noEmptyDestructuringRule.lint.ts
