+++ tests/rules/noIdenticalConditionsRule/noIdenticalConditionsRule.lint.ts
