+++ tests/rules/noIdenticalExpressionsRule/noIdenticalExpressionsRule.lint.ts
