+++ tests/rules/noIdenticalExpressionsRule/noIdenticalExpressionsRule.test.ts
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-import { Rule } from "../../../src/rules/NoIdenticalExpressionsRule";
+import { Rule } from "../../../src/rules/noIdenticalExpressionsRule";
 import runRule from "../../runRule";
 
 it("raises error", () => {
