+++ tests/rules/noIgnoredReturnRule/noIgnoredReturnRule.lint.ts
