+++ tests/rules/noSelfAssignmentRule/noSelfAssignmentRule.lint.ts
