+++ tests/rules/noUseOfEmptyReturnValueRule/noUseOfEmptyReturnValueRule.lint.ts
