+++ tests/rules/noVariableUsageBeforeDeclarationRule/noVariableUsageBeforeDeclarationRule.lint.ts
